Modularize SliderButton
